requirements.txt
alembic==1.6.5
Authlib==0.15.4
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
dnspython==1.16.0
ecdsa==0.17.0
eventlet==0.31.0
falcon==3.0.1
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
mongoengine==0.23.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pycparser==2.20
pymongo==3.11.4
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
python-jose==3.3.0
requests==2.25.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.19
urllib3==1.26.6
Werkzeug==2.0.1
